Многопоточность через наследника
